Fix cast warning - use size_t type when retrieving string indexes.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9480 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaAdapterLib/sipXmediaMediaProcessing/src/CpPhoneMediaInterface.cpp
@@ -1107,7 +1107,7 @@ OsStatus CpPhoneMediaInterface::playAudio(const char* url,
     OsStatus returnCode = OS_NOT_FOUND;
     UtlString urlString;
     if(url) urlString.append(url);
-    int fileIndex = urlString.index("file://");
+    size_t fileIndex = urlString.index("file://");
     if(fileIndex == 0) urlString.remove(0, 6);
 
     if(mpFlowGraph && !urlString.isNull())
